requirements.txt
requests~=2.32.3
Flask~=3.0.3
selenium~=4.21.0
gunicorn==20.1.0
Flask-Mail==0.9.1
flask_cors==3.0.10
webdriver_manager==3.5.2

dnspython~=2.6.1
msal~=1.28.0